Remove 'type: module' from the package.json file
This commit removes the "type: module" configuration from the package.json file. This configuration is used to indicate that a module is using ES6 module syntax instead of CommonJS syntax in Node.js. By removing this configuration, the module will default to using CommonJS syntax.

This change may be necessary if the module was previously using ES6 module syntax but needs to be compatible with older versions of Node.js or other environments that do not support it. Additionally, removing this configuration may improve the module's compatibility with other tools and libraries that are designed to work with CommonJS modules.